AVX512 encoding change.
The old format had to specify the broadcast round size as B32, or B64, and what the valid width bit setting is for (W=0), (W=1), by comparing the mode, and width bit setting instead of what the width bit has to be set too.

This version of the code compares the SIMD mode to what the valid width bit setting is.
The valid width bit size and Broadcast round size work together.

This version has been tested. It flawlessly decodes all AVX512 instructions like the original.

Both versions have pros, and cons.