AVX512 Encoding enhancement.
The old format had to specify the broadcast round size as B32, or B64, and what the valid width bit setting is for (W=0), (W=1), by comparing the width bit setting.

This version of the code compares the SIMD mode to what the valid width bit setting is. The valid width bit size, and Broadcast round size work together.

Some AVX512 instructions conflict with the correct SIMD mode setting for the width bit, and is an DWORD conversion over an Double 64 SIMD mode. These instructions require an extra separation in the opcode map in this version.

This code has been tested and flawlessly decodes all AVX512 instructions the same as the original, however both versions have pros, and cons to consider.

Pros.

1. Fewer comparisons to decoding an AVX512 instruction.
2. Slightly faster decoding for the majority of vector instructions.
3. Code is not much different, or bigger than the original.

Cons.

1. Uses 17% more Memory vs separating the (W=0), (W=1), and B32, B64 settings.
2. More steps to decoding some AVX512 instructions in the opcode map. Which can become progressively worse if more AVX512 instructions are added to future cores.